backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.2
psycopg[binary]>=3.1
dj-database-url>=2.1
python-dotenv>=1.0
channels[daphne]>=4.0
channels-redis>=4.2
Pillow>=10.0
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
drf-spectacular-sidecar>=2024.7.1
django-jazzmin>=3.0
